OCR/requirements.txt
appdirs==1.4.4
attrs==21.2.0
autopep8==1.5.7
black==19.3b0
Click==8.0.1
Pillow==8.2.0
pycodestyle==2.7.0
pytesseract==0.3.7
toml==0.10.2
